+++ VariantFamily/registration.php
@@ -1,0 +1,7 @@
+<?php
+
+\Magento\Framework\Component\ComponentRegistrar::register(
+    \Magento\Framework\Component\ComponentRegistrar::MODULE,
+    'Pimgento_VariantFamily',
+    __DIR__
+);